pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "colloquip"
version = "0.1.0"
description = "Emergent multi-agent deliberation system"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = [
    "anthropic>=0.40",
]
cli = [
    "rich>=13.0",
]
api = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "httpx>=0.27",
    "websockets>=13.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.8",
]

[project.scripts]
colloquip = "colloquip.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/colloquip"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: tests requiring real LLM API calls",
    "integration: full deliberation loop tests",
]

[tool.ruff]
line-length = 100
target-version = "py311"
